Pulled secondary component and util out of Conversation file, to reduce size
