+++ src/v2/Apps/Conversation/Components/Conversation.tsx
@@ -1,15 +1,5 @@
-import {
-  Box,
-  Flex,
-  Image,
-  Link,
-  Sans,
-  Spacer,
-  Spinner,
-  color,
-} from "@artsy/palette"
+import { Box, Flex, Spacer, Spinner } from "@artsy/palette"
 import { Conversation_conversation } from "v2/__generated__/Conversation_conversation.graphql"
-import { DateTime } from "luxon"
 import React, { useEffect, useRef, useState } from "react"
 import {
   RelayProp,
@@ -17,132 +7,13 @@ import {
   createPaginationContainer,
 } from "react-relay"
 import { graphql } from "relay-runtime"
+import { Item } from "./Item"
 import { MessageFragmentContainer as Message } from "./Message"
 import { Reply } from "./Reply"
 import { TimeSince, fromToday } from "./TimeSince"
 import { UpdateConversation } from "../Mutation/UpdateConversationMutation"
+import { groupMessages } from "../Utils/groupMessages"
 import styled from "styled-components"
-
-interface ItemProps {
-  item: Conversation_conversation["items"][0]["item"]
-}
-
-export type ItemType = "Artwork" | "Show"
-
-export const Item: React.FC<ItemProps> = props => {
-  const { item } = props
-  if (item.__typename === "%other") return null
-  const itemType = item.__typename as ItemType
-
-  const getImage = item => {
-    if (itemType === "Artwork") {
-      return item.image?.url
-    } else if (itemType === "Show") {
-      return item.coverImage?.url
-    } else {
-      return null
-    }
-  }
-
-  const itemDetails = item => {
-    if (item.__typename === "Artwork") {
-      return [
-        <Sans key={1} size="4" weight="medium" color="white100">
-          {item.artistNames}
-        </Sans>,
-        <Sans key={2} size="2" color="white100">
-          {item.title} / {item.date}
-        </Sans>,
-        item.listPrice?.display && (
-          <Sans key={3} size="2" color="white100">
-            {item.listPrice?.display}
-          </Sans>
-        ),
-      ]
-    } else if (item.__typename === "Show") {
-      const itemLocation = item.fair?.location?.city
-      return [
-        <Sans key={1} size="4" weight="medium" color="white100">
-          {item.fair.name}
-        </Sans>,
-        <Sans key={2} size="2" color="white100">
-          {itemLocation && `${itemLocation}, `}
-          {item?.fair?.exhibitionPeriod}
-        </Sans>,
-      ]
-    }
-  }
-
-  if (itemType === "Artwork" || itemType === "Show") {
-    return (
-      <Link
-        href={item.href}
-        underlineBehavior="none"
-        style={{ alignSelf: "flex-end" }}
-        mb={1}
-      >
-        <Flex flexDirection="column">
-          <Image
-            src={getImage(item)}
-            alt={item.__typename === "Artwork" ? item.title : item.name}
-            width="350px"
-            borderRadius="15px 15px 0 0"
-          />
-          <Flex
-            p={1}
-            flexDirection="column"
-            justifyContent="center"
-            background={color("black100")}
-            borderRadius="0 0 15px 15px"
-          >
-            {itemDetails(item)}
-          </Flex>
-        </Flex>
-      </Link>
-    )
-  } else {
-    return null
-  }
-}
-
-type Message = Conversation_conversation["messagesConnection"]["edges"][number]["node"]
-/**
- * Combines messages into groups of messages sent by the same party and
- * separated out into different groups if sent across multiple days
- * @param messages Messages in the conversation
- */
-const groupMessages = (messages: Message[]): Message[][] => {
-  if (messages.length === 0) {
-    return []
-  }
-  // Make a copy of messages
-  const remainingMessages = [...messages]
-  const groups = [[remainingMessages.pop()]]
-  while (remainingMessages.length > 0) {
-    const lastGroup = groups[groups.length - 1]
-    const lastMessage = lastGroup[lastGroup.length - 1]
-    const currentMessage = remainingMessages.pop()
-
-    const lastMessageCreatedAt = DateTime.fromISO(lastMessage.createdAt)
-    const currentMessageCreatedAt = DateTime.fromISO(currentMessage.createdAt)
-    const sameDay = lastMessageCreatedAt.hasSame(currentMessageCreatedAt, "day")
-
-    const today = fromToday(currentMessageCreatedAt)
-
-    if (sameDay && !today) {
-      lastGroup.push(currentMessage)
-    } else if (!today) {
-      groups.push([currentMessage])
-    } else if (lastMessage.isFromUser !== currentMessage.isFromUser) {
-      groups.push([currentMessage])
-    } else if (!sameDay && today) {
-      groups.push([currentMessage])
-    } else {
-      lastGroup.push(currentMessage)
-    }
-  }
-  return groups
-}
 
 export interface ConversationProps {
   conversation: Conversation_conversation
