+++ src/v2/Apps/Conversation/Utils/groupMessages.tsx
@@ -1,0 +1,42 @@
+import { Conversation_conversation } from "v2/__generated__/Conversation_conversation.graphql"
+import { DateTime } from "luxon"
+import { fromToday } from "../Components/TimeSince"
+
+type Message = Conversation_conversation["messagesConnection"]["edges"][number]["node"]
+/**
+ * Combines messages into groups of messages sent by the same party and
+ * separated out into different groups if sent across multiple days
+ * @param messages Messages in the conversation
+ */
+export const groupMessages = (messages: Message[]): Message[][] => {
+  if (messages.length === 0) {
+    return []
+  }
+  // Make a copy of messages
+  const remainingMessages = [...messages]
+  const groups = [[remainingMessages.pop()]]
+  while (remainingMessages.length > 0) {
+    const lastGroup = groups[groups.length - 1]
+    const lastMessage = lastGroup[lastGroup.length - 1]
+    const currentMessage = remainingMessages.pop()
+
+    const lastMessageCreatedAt = DateTime.fromISO(lastMessage.createdAt)
+    const currentMessageCreatedAt = DateTime.fromISO(currentMessage.createdAt)
+    const sameDay = lastMessageCreatedAt.hasSame(currentMessageCreatedAt, "day")
+
+    const today = fromToday(currentMessageCreatedAt)
+
+    if (sameDay && !today) {
+      lastGroup.push(currentMessage)
+    } else if (!today) {
+      groups.push([currentMessage])
+    } else if (lastMessage.isFromUser !== currentMessage.isFromUser) {
+      groups.push([currentMessage])
+    } else if (!sameDay && today) {
+      groups.push([currentMessage])
+    } else {
+      lastGroup.push(currentMessage)
+    }
+  }
+  return groups
+}
